+++ src/Android/Xamarin.Android/groups.json
@@ -130,8 +130,8 @@
               "Path": "Samples/Layers/ExportTiles"
             },
             {
-              "SampleName": "WebTiledLayer",
-              "Path": "Samples/Layers/WebTiledLayer" 
+              "SampleName": "Web_TiledLayer",
+              "Path": "Samples/Layers/Web_TiledLayer" 
             }
           ]
         },
