+++ src/Forms/Shared/Samples/Layers/Web_TiledLayer/Web_TiledLayer.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"             
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Xamarin.Forms;assembly=Esri.ArcGISRuntime.Xamarin.Forms"
              xmlns:mapping="clr-namespace:Esri.ArcGISRuntime.Mapping;assembly=Esri.ArcGISRuntime" 
-             x:Class="ArcGISRuntimeXamarin.Samples.WebTiledLayer.WebTiledLayer">
+             x:Class="ArcGISRuntimeXamarin.Samples.Web_TiledLayer.Web_TiledLayer">
   <Grid>
     <esriUI:MapView x:Name="MyMapView"/>
   </Grid>
