+++ src/UWP/ArcGISRuntime.UWP.Samples/Samples/Layers/Web_TiledLayer/Web_TiledLayer.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl 
-    x:Class="ArcGISRuntime.UWP.Samples.WebTiledLayer.WebTiledLayer"
+    x:Class="ArcGISRuntime.UWP.Samples.Web_TiledLayer.Web_TiledLayer"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
