--- src/WPF/ArcGISRuntime.WPF.Samples/Samples/Layers/WebTiledLayer/metadata.json
@@ -1,14 +1,0 @@
-{
-  "Name": "Web TiledLayer",
-  "SampleName": "WebTiledLayer",
-  "Description": "",
-  "Instructions": "", 
-  "Type": 0,
-  "RequiresOnlineConnection": true,
-  "RequiresOfflineData": false,
-  "RequiresLocalServer": false,
-  "Image": "WebTiledLayer.jpg",
-  "Link": "",
-  "TypeLink": [ ],
-  "SampleFolder": "WebTiledLayer"
-}
