+++ src/WPF/ArcGISRuntime.WPF.Samples/Samples/Layers/Web_TiledLayer/Web_TiledLayer.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.WebTiledLayer.WebTiledLayer"
+﻿<UserControl x:Class="ArcGISRuntime.WPF.Samples.Web_TiledLayer.Web_TiledLayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
