+++ src/iOS/Xamarin.iOS/Samples/Layers/Web_TiledLayer/Web_TiledLayer.cs
@@ -1,4 +1,4 @@
-// Copyright 2018 Esri.
+﻿// Copyright 2018 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
@@ -7,50 +7,52 @@
 // "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
 // language governing permissions and limitations under the License.
 
-using Android.App;
-using Android.OS;
-using Android.Widget;
 using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.UI.Controls;
+using Foundation;
+using UIKit;
 
-namespace ArcGISRuntimeXamarin.Samples.WebTiledLayer
+namespace ArcGISRuntimeXamarin.Samples.Web_TiledLayer
 {
-    [Activity]
-    public class WebTiledLayer : Activity
+    [Register("Web_TiledLayer")]
+    public class Web_TiledLayer : UIViewController
     {
         // Create and hold reference to the used MapView
         private MapView _myMapView = new MapView();
 
-        protected override void OnCreate(Bundle bundle)
+        public Web_TiledLayer()
         {
-            base.OnCreate(bundle);
-
             Title = "Web TiledLayer";
-
-            // Create the UI, setup the control references and execute initialization 
-            CreateLayout();
-            Initialize();
         }
 
         private void Initialize()
         {
             // Create new Map with basemap
             Map myMap = new Map(Basemap.CreateImagery());
-            
+
             // Provide used Map to the MapView
             _myMapView.Map = myMap;
         }
 
         private void CreateLayout()
         {
-            // Create a new vertical layout for the app
-            var layout = new LinearLayout(this) { Orientation = Orientation.Vertical };
+            // Add MapView to the page
+            View.AddSubviews(_myMapView);
+        }
 
-            // Add the map view to the layout
-            layout.AddView(_myMapView);
+        public override void ViewDidLoad()
+        {
+            CreateLayout();
+            Initialize();
 
-            // Show the layout in the app
-            SetContentView(layout);
+            base.ViewDidLoad();
+        }
+        public override void ViewDidLayoutSubviews()
+        {
+            // Setup the visual frame for the MapView
+            _myMapView.Frame = new CoreGraphics.CGRect(0, 0, View.Bounds.Width, View.Bounds.Height);
+
+            base.ViewDidLayoutSubviews();
         }
     }
 }
