+++ src/iOS/Xamarin.iOS/groups.json
@@ -131,8 +131,8 @@
               "Path": "Samples/Layers/ExportTiles"
             },
             {
-              "SampleName": "WebTiledLayer",
-              "Path": "Samples/Layers/WebTiledLayer" 
+              "SampleName": "Web_TiledLayer",
+              "Path": "Samples/Layers/Web_TiledLayer" 
             }
           ]
         },
